+++ test/0600-th_area_command_test.dart
@@ -53,7 +53,7 @@ scrap test
   area clay -clip on
     l85-3732--20
   endarea
-  line border -close on -id l85-3732--20 -visibility off
+  line border -id l85-3732--20 -close on -visibility off
     3592 208
     3539.45 249.03 3447.39 245.1 3392 208
     3233.22 101.65 3066.45 -131.93 3204 -332
@@ -88,7 +88,7 @@ scrap test
   area clay
     l85-3732--20
   endarea
-  line border -close on -id l85-3732--20 -visibility off
+  line border -id l85-3732--20 -close on -visibility off
     3592 208
     3539.45 249.03 3447.39 245.1 3392 208
     3233.22 101.65 3066.45 -131.93 3204 -332
@@ -111,7 +111,7 @@ scrap test
   area clay
     l85-3732--20
   endarea
-  line border -close on -id l85-3732--20 -visibility off
+  line border -id l85-3732--20 -close on -visibility off
     3592 208
     3539.45 249.03 3447.39 245.1 3392 208
     3233.22 101.65 3066.45 -131.93 3204 -332
@@ -130,7 +130,7 @@ scrap test
   area clay
     l85-3732--20
   endarea
-  line border -close on -id l85-3732--20 -visibility off
+  line border -id l85-3732--20 -close on -visibility off
     3592 208
     3539.45 249.03 3447.39 245.1 3392 208
     3233.22 101.65 3066.45 -131.93 3204 -332
@@ -174,7 +174,7 @@ scrap test
   area pool -clip on
     l85-3732--20
   endarea
-  line border -close on -id l85-3732--20 -visibility off
+  line border -id l85-3732--20 -close on -visibility off
     3592 208
     3539.45 249.03 3447.39 245.1 3392 208
     3233.22 101.65 3066.45 -131.93 3204 -332
@@ -219,7 +219,7 @@ scrap test
   area clay -clip off
     l85-3732--20
   endarea
-  line border -close on -id l85-3732--20 -visibility off
+  line border -id l85-3732--20 -close on -visibility off
     3592 208
     3539.45 249.03 3447.39 245.1 3392 208
     3233.22 101.65 3066.45 -131.93 3204 -332
@@ -238,7 +238,7 @@ scrap test
   area clay -clip off
     l85-3732--20
   endarea
-  line border -close on -id l85-3732--20 -visibility off
+  line border -id l85-3732--20 -close on -visibility off
     3592 208
     3539.45 249.03 3447.39 245.1 3392 208
     3233.22 101.65 3066.45 -131.93 3204 -332
@@ -307,7 +307,7 @@ scrap test
   area water -context point anchor
     l85-3732--20
   endarea
-  line border -close on -id l85-3732--20 -visibility off
+  line border -id l85-3732--20 -close on -visibility off
     3592 208
     3539.45 249.03 3447.39 245.1 3392 208
     3233.22 101.65 3066.45 -131.93 3204 -332
@@ -327,7 +327,7 @@ scrap test
   area water -context point anchor
     l85-3732--20
   endarea
-  line border -close on -id l85-3732--20 -visibility off
+  line border -id l85-3732--20 -close on -visibility off
     3592 208
     3539.45 249.03 3447.39 245.1 3392 208
     3233.22 101.65 3066.45 -131.93 3204 -332
@@ -369,7 +369,7 @@ scrap test
   area water -id area51
     l85-3732--20
   endarea
-  line border -close on -id l85-3732--20 -visibility off
+  line border -id l85-3732--20 -close on -visibility off
     3592 208
     3539.45 249.03 3447.39 245.1 3392 208
     3233.22 101.65 3066.45 -131.93 3204 -332
@@ -388,7 +388,7 @@ scrap test
   area water -id area51
     l85-3732--20
   endarea
-  line border -close on -id l85-3732--20 -visibility off
+  line border -id l85-3732--20 -close on -visibility off
     3592 208
     3539.45 249.03 3447.39 245.1 3392 208
     3233.22 101.65 3066.45 -131.93 3204 -332
@@ -430,7 +430,7 @@ scrap test
   area clay -place top
     l85-3732--20
   endarea
-  line border -close on -id l85-3732--20 -visibility off
+  line border -id l85-3732--20 -close on -visibility off
     3592 208
     3539.45 249.03 3447.39 245.1 3392 208
     3233.22 101.65 3066.45 -131.93 3204 -332
@@ -449,7 +449,7 @@ scrap test
   area clay -place top
     l85-3732--20
   endarea
-  line border -close on -id l85-3732--20 -visibility off
+  line border -id l85-3732--20 -close on -visibility off
     3592 208
     3539.45 249.03 3447.39 245.1 3392 208
     3233.22 101.65 3066.45 -131.93 3204 -332
@@ -469,7 +469,7 @@ scrap test
   area clay -place top # With a same line comment in the command like option
     l85-3732--20
   endarea
-  line border -close on -id l85-3732--20 -visibility off
+  line border -id l85-3732--20 -close on -visibility off
     3592 208
     3539.45 249.03 3447.39 245.1 3392 208
     3233.22 101.65 3066.45 -131.93 3204 -332
@@ -538,7 +538,7 @@ scrap test
   area clay -visibility off
     l85-3732--20
   endarea
-  line border -close on -id l85-3732--20 -visibility off
+  line border -id l85-3732--20 -close on -visibility off
     3592 208
     3539.45 249.03 3447.39 245.1 3392 208
     3233.22 101.65 3066.45 -131.93 3204 -332
@@ -558,7 +558,7 @@ scrap test
   area clay -visibility off
     l85-3732--20
   endarea
-  line border -close on -id l85-3732--20 -visibility off
+  line border -id l85-3732--20 -close on -visibility off
     3592 208
     3539.45 249.03 3447.39 245.1 3392 208
     3233.22 101.65 3066.45 -131.93 3204 -332